### 2020-07-13 77771f8
Update GPU_kernels_single_precision.cu
