+++ GENRE_Code/GENRE_GPU_Single_Precision_Code/GPU_kernels_single_precision.cu
@@ -506,8 +506,8 @@ if (block_thread_ind < num_threads_per_block_2) {
 
       // Obtain the flag that determines whether the first predictor column is a column of ones for the intercept term or not
       int intercept_flag = (int)intercept_flag_d[fit_ind];
-
-      // Declare and initialize the variable that stores the maximum predictor coefficient change
+           
+      // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
       float global_max_change = 1E12;
 
       // Declare and initialize the variable that counts how many iterations of cyclic coordinate descent have been performed
@@ -519,12 +519,12 @@ if (block_thread_ind < num_threads_per_block_2) {
           sdata[store_ind] = residual_y_d[observation_thread_stride + observation_row];
       }
 
-      // Perform cyclic coordinate descent until either the maximum number of iterations convergence criterion or the tolerance criterion is met 
+      // Perform cyclic coordinate descent until either the maximum number of iterations is reached or the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values becomes less than the tolerance
       while (global_max_change >= tolerance && iteration_count < max_iterations) {
-            // Declare and initialize the variable that stores the maximum predictor coefficient change for one iteration
+            // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
             float max_change = 0.0f;
 
-            // Declare and initialize the variable that stores the change between the current predictor coefficient value and its previous value
+            // Declare and initialize the variable that stores the weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values that are due to the current predictor coefficient value being updated using cyclic coordinate descent
             float change = 0.0f;
          
             // Cycle through all of the predictors for one iteration of cyclic coordinate descent
