### 2020-07-13 4259a02
Update GPU_kernels_double_precision.cu
