+++ README.md
@@ -70,7 +70,7 @@ fits will be saved to.
 
 ```lambda_values_h```: A vector containing ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. Note that ```GENRE``` only computes the model coefficients for one value of ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. This is different from other packages like ```glmnet```, which compute the coefficients for a path of lambda values.
 
-```tolerance_values_h```: A vector containing the tolerance convergence criterion value for each model fit (values such as 1E-4 or 1E-5 are reasonable). ```GENRE``` uses the same tolerance convergence criterion as ```glmnet```. Each time a model coefficient is updated, the weighted (all of the observation weights are 1 in this case) sum of squares of the changes of the fitted values due to this update is calculated as ![tolerance_convergence_criterion](https://latex.codecogs.com/svg.latex?%5Cinline%20%5Cfrac%7B1%7D%7BN%7D%5Csum_%7Bi%3D1%7D%5E%7BN%7D%28%5Cboldsymbol%7BX%7D_%7Bij%7D%5Cboldsymbol%7B%5Chat%7B%5Cbeta%7D%7D_%7Bj%7D%5E%7B%28k%29%7D-%5Cboldsymbol%7BX%7D_%7Bij%7D%5Cboldsymbol%7B%5Chat%7B%5Cbeta%7D%7D_%7Bj%7D%5E%7B%28k&plus;1%29%7D%29%5E%7B2%7D), where ![beta_k](https://latex.codecogs.com/svg.latex?%5Cboldsymbol%7B%5Chat%7B%5Cbeta%7D%7D_%7Bj%7D%5E%7B%28k%29%7D) is the value of the model coefficient before the update and ![beta_k_plus_1](https://latex.codecogs.com/svg.latex?%5Cboldsymbol%7B%5Chat%7B%5Cbeta%7D%7D_%7Bj%7D%5E%7B%28k&plus;1%29%7D) is the value of the model coefficient after the update. This will be evaluated for every model coefficient until one iteration of cyclic coordinate descent is completed. The maximum value for this evaluated term is then taken across all of the model coefficients, and it is compared to the specified tolerance value. If the value is less than the tolerance value, then cyclic coordinate descent will stop. However, if it is equal to or greater than the tolerance value, then cyclic coordinate descent will continue.
[…]
