### 2020-08-02 795ace9
Update GPU_kernels_single_precision.cu
