+++ GENRE_Code/GENRE_GPU_Single_Precision_Code/GPU_kernels_single_precision.cu
@@ -258,7 +258,7 @@ if (block_thread_ind < num_threads_per_block_2) {
 
 
                          
-// Define the GPU kernel that performs least squares regression with elastic-net regularization using the cyclic coordinate descent optimization algorithm in order to fit the model matrices to the data
+// Define the GPU kernel that performs least-squares regression with elastic-net regularization using the cyclic coordinate descent optimization algorithm in order to fit the model matrices to the data
 __global__ void model_fit(float * B_d, double * B_thread_stride_d, float * model_fit_flag_d, float * X_matrix_d, double * X_matrix_thread_stride_d, double * observation_thread_stride_d, float * residual_y_d, float * y_std_d, float * standardized_lambda_values_d, double * num_observations_d, double * num_predictors_d, float * alpha_values_d, float * tolerance_values_d, float * max_iterations_values_d, float * intercept_flag_d, int transformation_flag, int num_threads_per_block, int num_blocks, int num_threads_last_block) {
 
 // Obtain the index of the block
@@ -313,18 +313,18 @@ if (block_thread_ind < num_threads_per_block_2) {
       // Obtain the flag that determines whether the first predictor column is a column of ones for the intercept term or not
       int intercept_flag = (int)intercept_flag_d[fit_ind];
 
-      // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
+      // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values for one iteration of cyclic coordinate descent
       float global_max_change = 1E12;
 
       // Declare and initialize the variable that counts how many iterations of cyclic coordinate descent have been performed
       int iteration_count = 0;
 
-      // Perform cyclic coordinate descent until either the maximum number of iterations is reached or the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values becomes less than the tolerance
+      // Perform cyclic coordinate descent until either the maximum number of iterations is reached or the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values becomes less than the tolerance
       while (global_max_change >= tolerance && iteration_count < max_iterations) {
-            // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
+            // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values for one iteration of cyclic coordinate descent
             float max_change = 0.0f;
 
-            // Declare and initialize the variable that stores the weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values that are due to the current predictor coefficient value being updated using cyclic coordinate descent
+            // Declare and initialize the variable that stores the weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values that are due to the current predictor coefficient value being updated using cyclic coordinate descent
             float change = 0.0f;
          
             // Cycle through all of the predictors for one iteration of cyclic coordinate descent
@@ -406,7 +406,7 @@ if (block_thread_ind < num_threads_per_block_2) {
                 B_d[predictor_thread_stride + j] = B_j;
 
                 // Update the residual values to include the contribution of the current predictor using the updated predictor coefficient value 
-                // If the updated predictor coefficient value is 0, then it's contribution to the residual values is zero
+                // If the updated predictor coefficient value is 0, then its contribution to the residual values is zero
                 if (B_j != 0.0f) {
[…]
