### 2020-08-02 9e99d02
Update GPU_kernels_double_precision.cu
