+++ GENRE_Code/GENRE_GPU_Double_Precision_Code/GPU_kernels_double_precision.cu
@@ -257,7 +257,7 @@ if (block_thread_ind < num_threads_per_block_2) {
 
 
 
-// Define the GPU kernel that performs least squares regression with elastic-net regularization using the cyclic coordinate descent optimization algorithm in order to fit the model matrices to the data
+// Define the GPU kernel that performs least-squares regression with elastic-net regularization using the cyclic coordinate descent optimization algorithm in order to fit the model matrices to the data
 __global__ void model_fit(double * B_d, double * B_thread_stride_d, double * model_fit_flag_d, double * X_matrix_d, double * X_matrix_thread_stride_d, double * observation_thread_stride_d, double * residual_y_d, double * y_std_d, double * standardized_lambda_values_d, double * num_observations_d, double * num_predictors_d, double * alpha_values_d, double * tolerance_values_d, double * max_iterations_values_d, double * intercept_flag_d, int transformation_flag, int num_threads_per_block, int num_blocks, int num_threads_last_block) {
 
 // Obtain the index of the block
@@ -312,18 +312,18 @@ if (block_thread_ind < num_threads_per_block_2) {
       // Obtain the flag that determines whether the first predictor column is a column of ones for the intercept term or not
       int intercept_flag = (int)intercept_flag_d[fit_ind];
 
-      // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
+      // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values for one iteration of cyclic coordinate descent
       double global_max_change = 1E12;
 
       // Declare and initialize the variable that counts how many iterations of cyclic coordinate descent have been performed
       int iteration_count = 0;
 
-      // Perform cyclic coordinate descent until either the maximum number of iterations is reached or the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values becomes less than the tolerance           
+      // Perform cyclic coordinate descent until either the maximum number of iterations is reached or the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values becomes less than the tolerance           
       while (global_max_change >= tolerance && iteration_count < max_iterations) {
-            // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values for one iteration of cyclic coordinate descent
+            // Declare and initialize the variable that stores the maximum weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values for one iteration of cyclic coordinate descent
             double max_change = 0.0;
 
-            // Declare and initialize the variable that stores the weighted (observation weights are all 1 in this case) sum of squares of the changes of the fitted values that are due to the current predictor coefficient value being updated using cyclic coordinate descent
+            // Declare and initialize the variable that stores the weighted (observation weights are all 1 in this case) sum of squares of the changes in the fitted values that are due to the current predictor coefficient value being updated using cyclic coordinate descent
             double change = 0.0;
          
             // Cycle through all of the predictors for one iteration of cyclic coordinate descent
@@ -372,7 +372,7 @@ if (block_thread_ind < num_threads_per_block_2) {
                 // Divide the computed correlation by the total number of observations in y (also the total number of observations in one predictor column)
                 p_j = ((double)1.0 / (double)num_observations) * p_j;
 
-                // Apply the soft-thresholding function that is associated with the L1 regularization component of elastic-net regularization 
+                // Apply the soft-thresholding function that is associated with the L1-regularization component of elastic-net regularization 
                 double gamma = lambda * alpha;
[…]
