### 2020-08-02 d06d0a4
Update GPU_kernels_single_precision.cu
