### 2020-08-02 fe9ebe9
Update GPU_kernels_single_precision.cu
