### 2020-08-02 a1475a1
Update GPU_kernels_double_precision.cu
