+++ GENRE_Code/GENRE_GPU_Double_Precision_Code/GPU_kernels_double_precision.cu
@@ -398,6 +398,7 @@ if (block_thread_ind < num_threads_per_block_2) {
                    B_j = p_j;
                 } else {
                    // Calculate the updated predictor coefficient value by applying the component of elastic-net regularization that is associated with L2-regularization 
+                   // The mean_squared_predictor_value term comes from the derivation of the coordinate descent update for a predictor coefficient         
                    B_j = B_j / (mean_squared_predictor_value + (lambda * ((double)1.0 - alpha)));
                 }
 
@@ -598,6 +599,7 @@ if (block_thread_ind < num_threads_per_block_2) {
                    B_j = p_j;
                 } else {
                    // Calculate the updated predictor coefficient value by applying the component of elastic-net regularization that is associated with L2-regularization 
+                   // The mean_squared_predictor_value term comes from the derivation of the coordinate descent update for a predictor coefficient         
                    B_j = B_j / (mean_squared_predictor_value + (lambda * ((double)1.0 - alpha)));
                 }
 
