### 2020-08-05 #2475 @sjpfenninger
@mlxd and @marouenbg would you be willing to review this submission for JOSS?
