### 2020-08-05 #2475 @mlxd
@sjpfenninger Sorry, but I do not have access to a MATLAB license on a CUDA-capable machine, and so would be unable to properly review this.
