### 2020-09-07 #2644 @whedon
[REVIEW]: GENRE (GPU Elastic-Net REgression): A CUDA-Accelerated Package for Massively Parallel Linear Regression with Elastic-Net Regularization
**Submitting author:** @Christopher-Khan (<a href="http://orcid.org/0000-0003-3201-3423">Christopher Khan</a>)
**Repository:** <a href="https://github.com/VU-BEAM-Lab/GENRE" target ="_blank">https://github.com/VU-BEAM-Lab/GENRE</a>
**Version:** v1.0.1
