### 2020-09-08 VU-BEAM-Lab/GENRE#1 @marouenbg
JOSS review #1
Hi @Christopher-Khan ,

I will post my reviews here as I go through the software
https://github.com/openjournals/joss-reviews/issues/2644
