### 2020-09-08 VU-BEAM-Lab/GENRE#1 @marouenbg
GENRE is a GPU implementation of elastic-net that allows to handle several regression problems in parallel. It is provided as a CUDA implementation with an interface to MATLAB through MEX that allows a larger user base to access the software. The package is well documented, it is implemented in a scripting language and a programming language.

Here are my initial comments:

- You mentioned that GENRE was twice faster in GPU than CPU. Can you add a table with the benchmark results?

- Consider adding community guidelines or a contribution guide, consider moving the `contribution` section in README to a `contribute.md`

- The packages needs a manual (a test function with a toy example) or an automatic testing component (Travis/ Github actions / Appveyor).

In the manuscript:

- The multiple references format is the following [@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)

- The full reference of MATLAB is (The MathWorks, Inc., Natick, MA, USA)

- In the statement of need section, please specify the acronym: Aperture Domain Model Image
REconstruction (ADMIRE)

- In 'However, a GPU implementation of the
algorithm was developed' I think you need to specify that you're talking about GENRE here

- extend the acronyms fMRI, CT, PET in words
