+++ Paper/paper.md
@@ -30,7 +30,7 @@ combination of L1-regularization and L2-regularization to be applied to a given 
 Therefore, feature selection and coefficient shrinkage are performed while still allowing for the 
 presence of groups of correlated features. The process of performing these model fits can be 
 computationally expensive, and one of the fastest packages that is currently available is glmnet 
-[@friedman_hastie_tibshirani_2010], [@qian_hastie_friedman_tibshirani_simon_2013], [@hastie_qian_2014]. 
+[@friedman_hastie_tibshirani_2010; @qian_hastie_friedman_tibshirani_simon_2013; @hastie_qian_2014]. 
 This package provides highly efficient Fortran implementations of several different types of regression. 
 In the case of its implementation of linear regression with elastic-net regularization, the objective 
 function shown in (eq. \ref{eq:1}) is minimized. 
@@ -53,9 +53,9 @@ another interface if it is desired to call the C/CUDA code in another language, 
 without an interface.
 
 # Statement of Need
-The core motivation for developing GENRE was that many of the available packages for performing linear regression with elastic-net regularization focus on achieving high performance in terms of computational time or resource consumption for single model fits. However, they often do not address the case in which there is a need to perform many model fits in parallel. For example, the research project that laid the foundation for GENRE involved performing ultrasound image reconstruction using an algorithm called Aperture Domain Model Image REconstruction [@byram_jakovljevic_2014], [@byram_dei_tierney_dumont_2015], [@dei_byram_2017]. This algorithm is computationally expensive due to the fact that in one stage, it requires thousands of instances of linear regression with elastic-net regularization to be performed in order to fit models of ultrasound data. Originally, this algorithm was implemented on a CPU, and it typically required several minutes to reconstruct one ultrasound image. The primary bottleneck was performing all of the required model fits due to the fact that glmnet was used to compute each fit serially. However, a GPU implementation of the algorithm was developed, and this implementation provided a speedup of over two orders of magnitude, which allowed for multiple ultrasound images to be reconstructed per second. The main contributor to this speedup was the fact that the model fits were performed in parallel on the GPU. 
+The core motivation for developing GENRE was that many of the available packages for performing linear regression with elastic-net regularization focus on achieving high performance in terms of computational time or resource consumption for single model fits. However, they often do not address the case in which there is a need to perform many model fits in parallel. For example, the research project that laid the foundation for GENRE involved performing ultrasound image reconstruction using an algorithm called Aperture Domain Model Image REconstruction (ADMIRE) [@byram_jakovljevic_2014; @byram_dei_tierney_dumont_2015; @dei_byram_2017]. This algorithm is computationally expensive due to the fact that in one stage, it requires thousands of instances of linear regression with elastic-net regularization to be performed in order to fit models of ultrasound data. Originally, this algorithm was implemented on a CPU, and it typically required several minutes to reconstruct one ultrasound image. The primary bottleneck was performing all of the required model fits due to the fact that glmnet was used to compute each fit serially. However, a GPU implementation of the algorithm was developed, and this implementation provided a speedup of over two orders of magnitude, which allowed for multiple ultrasound images to be reconstructed per second. The main contributor to this speedup was the fact that the model fits were performed in parallel on the GPU. 
 
[…]
