### 2020-09-08 c4e98c3
Update paper.md
