+++ Paper/paper.md
@@ -55,8 +55,7 @@ without an interface.
 # Statement of Need
 The core motivation for developing GENRE was that many of the available packages for performing linear regression with elastic-net regularization focus on achieving high performance in terms of computational time or resource consumption for single model fits. However, they often do not address the case in which there is a need to perform many model fits in parallel. For example, the research project that laid the foundation for GENRE involved performing ultrasound image reconstruction using an algorithm called Aperture Domain Model Image REconstruction (ADMIRE) [@byram_jakovljevic_2014; @byram_dei_tierney_dumont_2015; @dei_byram_2017]. This algorithm is computationally expensive due to the fact that in one stage, it requires thousands of instances of linear regression with elastic-net regularization to be performed in order to fit models of ultrasound data. Originally, this algorithm was implemented on a CPU, and it typically required several minutes to reconstruct one ultrasound image. The primary bottleneck was performing all of the required model fits due to the fact that glmnet was used to compute each fit serially. However, a GPU implementation of the algorithm was developed, and this implementation provided a speedup of over two orders of magnitude, which allowed for multiple ultrasound images to be reconstructed per second. The main contributor to this speedup was the fact that the model fits were performed in parallel on the GPU. 
 
-Aside from this application, there are a number of other applications that can potentially benefit from having the ability to perform model fits in a massively parallel fashion, which is why the code was developed into a package. For example, linear regression with elastic-net regularization has been applied to the field of genomics in order to develop predictive models that utilize genetic markers [@ogutu_schulz-streeck_piepho_2012; @waldmann_mészáros_gredler_fuerst_sölkner_2013]. In addition, like ADMIRE, there are a variety of other signal processing applications. For example, this regression method has been used to create models of fMRI data in order to predict the mental states of subjects and provide insight into neural activity [@carroll_cecchi_rish_garg_rao_2009]. Moreover, another signal processing example is that linear regression models with elastic-net regularization have been used in combination with hidden Markov random field segmentation to perform CT estimation for the purposes of MRI-based attenuation correction for PET/MRI
-[@chen_juttukonda_lee_su_espinoza_lin_shen_lulash_an_2014]. Now, by using GENRE, the models in each of the aforementioned examples can be computed in parallel in order to reduce the amount of processing time that is required.
+Aside from this application, there are a number of other applications that can potentially benefit from having the ability to perform model fits in a massively parallel fashion, which is why the code was developed into a package. For example, linear regression with elastic-net regularization has been applied to the field of genomics in order to develop predictive models that utilize genetic markers [@ogutu_schulz-streeck_piepho_2012; @waldmann_mészáros_gredler_fuerst_sölkner_2013]. In addition, like ADMIRE, there are a variety of other signal processing applications. For example, this regression method has been used to create models of functional magnetic resonance imaging data in order to predict the mental states of subjects and provide insight into neural activity [@carroll_cecchi_rish_garg_rao_2009]. Moreover, another signal processing example is that linear regression models with elastic-net regularization have been used in combination with hidden Markov random field segmentation to perform computed tomography estimation for the purposes of magnetic resonance imaging-based attenuation correction for positron emission tomography/magnetic resonance imaging [@chen_juttukonda_lee_su_espinoza_lin_shen_lulash_an_2014]. Now, by using GENRE, the models in each of the aforementioned examples can be computed in parallel in order to reduce the amount of processing time that is required.
 
 # Acknowledgements
 This work was supported by NIH grants R01EB020040 and S10OD016216-01 and NAVSEA grant N0002419C4302.
