### 2020-09-09 VU-BEAM-Lab/GENRE#1 @marouenbg
The README file is quite complete, I had a couple of comments about the format.

- In the README `setup` section, please make the installation instructions as ordered bullet points with a specific header for Windows and Linux rather than a paragraph. You can add a troubleshooting step for each bullet point

- same for `model data format` section, consider having bullet points for the variables and the files

- transformation flag values in `user-defined parameters` also could be in bullet points

- I think that README and tutorial are quite informative, consider transforming them into a MATLAB livescript if you think this could be useful.

Going back to my first comment about benchmarking results, how do you think GENRE would perform in comparison to a multi-CPU LASSO implementation? I am interested in the effect of clock speed in CPUs vs GPUs.
