### 2020-09-10 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
Hi @marouenbg! Thank you for taking the time to review my submission and for your helpful comments. I really appreciate it. I have addressed your comments about the manuscript and will continue to address the remaining comments. I was a little slow to respond because I've been attending a virtual conference this week. I just wanted to ask about two of the above comments.

1. For the comment that talks about GENRE being two times faster than a CPU, is this referring to the section in my paper where I say that the GPU implementation of ADMIRE was two orders of magnitude faster than the CPU implementation? I can post the benchmarks, but I was able to optimize the GPU elastic-net even more in that specific case. For example, groups of model fits could use the same model matrix in that case, which allowed me to achieve improved coalesced memory access and to better utilize GPU memory bandwidth. Essentially, I had developed the GPU elastic-net for our ADMIRE project and then decided to make it a generalized software package (GENRE). As a result of this generalization, I just assumed each model fit uses a unique model matrix. Therefore, did you still want me to include the ADMIRE-specific benchmarks. I was thinking it might be better to include a benchmark comparing GENRE to a serial C implementation because it can still provide significant speedup. In terms of how it would compare to a multi-CPU implementation, I haven't benchmarked that (I haven't used OpenMP or MPI to create a multi-CPU implementation), but I think it would be possible to get similar speedup depending on how many CPU cores are available and the resources that are available for each core. However, a user could take the GPU kernels from GENRE and optimize them even more for their specific application like I did for ADMIRE, which is why I tried to extensively comment the code :). For example, in the case of ADMIRE, optimizing it even more made the difference of going from one order of magnitude speedup to two orders of magnitude when compared to a serial CPU implementation. In a case like this, it will be more difficult for a multi-CPU implementation to achieve similar performance. Did you have a specific multi-CPU package in mind? I ask because many of the papers I've seen usually focus on using multiple CPUs for just one model fit and not multiple in parallel.

2. In regards to the toy model, I do have a tutorial in the README, which involves generating toy data and processing it using GENRE. However, I can add more if it is not sufficient.

Thank you again for reviewing this submission.
