+++ README.md
@@ -43,6 +43,7 @@ The description provided above describes the process of performing one model fit
 
 ### Code Compilation for Windows OS
 * Go to the MATLAB directory that contains the repository folders and add them to your MATLAB path. Following this, go to the ```GENRE_Code``` folder. Type the following commands into the MATLAB command prompt.
+  * The ```-v``` flag can also be included at the end of each mexcuda command to display compilation details. If the compilation process is successful, then it will display a success message for each compilation in the command prompt. In addition, a compiled MEX-file will appear in each folder. The compilation process is important, and it is recommended to recompile any time a different release of MATLAB is utilized.
 
 ```Matlab
 cd GENRE_GPU_Single_Precision_Code
@@ -50,8 +51,6 @@ mexcuda GENRE_GPU_single_precision.cu
 cd ..\GENRE_GPU_Double_Precision_Code
 mexcuda GENRE_GPU_double_precision.cu
 ```
-
-The ```-v``` flag can also be included at the end of each mexcuda command to display compilation details. If the compilation process is successful, then it will display a success message for each compilation in the command prompt. In addition, a compiled MEX-file will appear in each folder. The compilation process is important, and it is recommended to recompile any time a different release of MATLAB is utilized.
 
 ### Code Compilation for Linux OS
 * Go to the MATLAB directory that contains the repository folders and add them to your MATLAB path. Following this, go to the ```GENRE_Code``` folder. Type the following commands into the MATLAB command prompt. 
