+++ README.md
@@ -286,10 +286,9 @@ Note that since we included an intercept term in every model, the first model co
    require less memory because multiple copies of the model will not be required.
    
 5. When possible, ```GENRE``` uses shared memory on the GPU in addition to global memory when performing the model fits. This memory has lower latency than global memory, so it can 
-   potentially improve performance. Whether shared memory is utilized or not is determined in the ```GENRE.m``` 
-   script, but it will usually be used if for the model matrix with the largest number of observations, the number of observations is less 
-   than or equal to 250 observations for single precision and less than or equal to 125 observations for double precision. Note that if this 
-   requirement is not met, then GPU global memory will be used for the model fits.
+   potentially improve performance. Whether shared memory is utilized or not is determined in the ```GENRE.m``` script, but it will usually be used if there is at least 32 KB of 
+   shared memory per GPU block and if for the model matrix with the largest number of observations, the number of observations is less than or equal to 250 observations for single 
+   precision and less than or equal to 125 observations for double precision. Note that if these two requirements are not met, then shared memory will not be used for the model fits.
    
 ## Comparing with Other Packages
 If you want to compare ```GENRE``` to other packages that perform linear regression with elastic-net regularization, you should check to see what convergence criteria these packages use in order to ensure that convergence for ```GENRE``` and the package being compared to it is being reached at similar points during optimization. For example, ```GENRE``` was developed to produce similar results to ```glmnet```, and both packages use the same tolerance convergence criterion. Therefore, if you are comparing these two packages, make sure to set the ```thresh``` parameter of each model fit in ```glmnet``` to be the same as the corresponding value in the ```tolerance_values_h``` vector of ```GENRE```. The code for the MATLAB port of ```glmnet``` is provided at http://web.stanford.edu/~hastie/glmnet_matlab/index.html. The numerical precision that is being used by the package should also match whatever option is selected for the ```precision``` parameter in ```GENRE```. In addition, when comparing with any package, make sure to use similar predictor transformations in regards to normalization or standardization. Adding on, for timing purposes, note that the ```GENRE.m``` script also includes loading in and organizing the data for all of the model fits because the ```data_organizer.m``` script is called within this script. It also includes saving the parameters and the computed model coefficients for the model fits. Moreover, as discussed in the [Additional Notes](#Additional-Notes) section, there is additional overhead when calling either of the MEX-files for the first time after ```clear mex``` has been called. As a result, these factors should be noted when comparing the time that it takes for ```GENRE``` to run to the time that it takes for another package to run. In terms of speedup, the users who will most likely experience the greatest speedup using ```GENRE``` when compared to other packages are those that need to perform many small fits and analyses such as trying out many different combinations of ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) and ![alpha](https://latex.codecogs.com/svg.latex?%5Calpha).
