+++ README.md
@@ -77,30 +77,30 @@ All of the ```.mat``` files should be saved in a directory. In terms of the nami
 ## User-Defined Parameters
 ```GENRE``` consists of several files. The main program is the ```GENRE.m``` script, and it is the only file that the user will need to modify. The inputs to this script are described in detail below.
 
-```precision```: Specifies which numerical precision to use for the model fit calculations on the GPU. The two options are either ```precision = 'single'``` or ```precision = 'double'```. Using double precision instead of single precision on GPUs typically results in a performance penalty due to there being fewer FP64 units than FP32 units and double precision requiring more memory resources as a result of one value of type double being 64 bits versus one value of type single being 32 bits. However, using single precision has the trade-off of reduced numerical precision. Depending on factors such as the conditioning of the model matrices, this reduced precision can lead to significantly different results. Therefore, if you select single precision, then you should ensure that this precision is sufficient for your application. If you are uncertain, then it is recommended to use double precision.
+* ```precision```: Specifies which numerical precision to use for the model fit calculations on the GPU. The two options are either ```precision = 'single'``` or ```precision = 'double'```. Using double precision instead of single precision on GPUs typically results in a performance penalty due to there being fewer FP64 units than FP32 units and double precision requiring more memory resources as a result of one value of type double being 64 bits versus one value of type single being 32 bits. However, using single precision has the trade-off of reduced numerical precision. Depending on factors such as the conditioning of the model matrices, this reduced precision can lead to significantly different results. Therefore, if you select single precision, then you should ensure that this precision is sufficient for your application. If you are uncertain, then it is recommended to use double precision.
 
-```num_fits```: The number of model fits to perform.
+* ```num_fits```: The number of model fits to perform.
 
-```data_path```: Path to the directory containing the model data files described in the previous section.
+* ```data_path```: Path to the directory containing the model data files described in the previous section.
 
-```save_path```: Path to the directory where the output file containing the parameters and the computed model coefficients for the model 
+* ```save_path```: Path to the directory where the output file containing the parameters and the computed model coefficients for the model 
 fits will be saved to.
 
-```output_filename```: The name of the output file containing the parameters and the computed model coefficients for the model fits.
+* ```output_filename```: The name of the output file containing the parameters and the computed model coefficients for the model fits.
 
-```alpha_values_h```: A vector containing ![alpha](https://latex.codecogs.com/svg.latex?%5Calpha) for each model fit.
+* ```alpha_values_h```: A vector containing ![alpha](https://latex.codecogs.com/svg.latex?%5Calpha) for each model fit.
 
-```lambda_values_h```: A vector containing ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. Note that ```GENRE``` only computes the model coefficients for one value of ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. This is different from other packages like ```glmnet```, which compute the coefficients for a path of lambda values.
+* ```lambda_values_h```: A vector containing ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. Note that ```GENRE``` only computes the model coefficients for one value of ![lambda](https://latex.codecogs.com/svg.latex?%5Clambda) for each model fit. This is different from other packages like ```glmnet```, which compute the coefficients for a path of lambda values.
 
[…]
