+++ README.md
@@ -303,7 +303,7 @@ Copyright 2020 Christopher Khan
 If you have questions, suggestions, or bugs to report, please use the Issues section of this repository or contact Christopher Khan (christopher.m.khan@vanderbilt.edu). 
 
 ## Contributing
-For contributing to this repository, please fork the repository and submit your proposed contributions in the form of a pull request. When submitting a pull request, please provide a detailed description of the proposed contributions. The pull request will then be reviewed by the repository manager ![@Christopher-Khan]. If it is approved, the pull request will be merged with the master branch of the repository. 
+For contributing to this repository, please fork the repository and submit your proposed contributions in the form of a pull request. When submitting a pull request, please provide a detailed description of the proposed contributions. The pull request will then be reviewed by the repository manager @Christopher-Khan. If it is approved, the pull request will be merged with the master branch of the repository. 
 
 ## Acknowledgements
 This work was supported by NIH grants R01EB020040 and S10OD016216-01 and NAVSEA grant N0002419C4302.
