### 2020-09-11 VU-BEAM-Lab/GENRE#1 @marouenbg
Hi @Christopher-Khan ,
Thanks for the feedback, I had fun reviewing this quality contribution.

Regarding question 1, yes if you can post the benchmark where GENRE is twice as fast in comparison to CPU and also mention the ADMIRE specific contribution, that would support the central claim in this paper.
About multi-CPU implementation, I think your answer above could be added in the discussion section and that would be enough in my opinion, there is no need to add more benchmarks.

Regarding question 2, I meant that you can use the toy model to implement a test functionality using [TRAVIS-CI](https://travis-ci.org/) for example. If you're not familiar with continuous integration, you can write a series of small tests (which can be based on your tutorial) and use the `assert` function to determine that software and hardware are properly configured. You can call the function `testGENRE()` and it could be a quick way for users to tell if everything is working properly.
