### 2020-09-11 VU-BEAM-Lab/GENRE#1 @marouenbg
Continuing with the review:

- I run the tutorial on a K80 machine, AWS p2.xlarge. I think `GENRE.m` should be a function that can be called as `GENRE()` or `GENRE_Tutorial()` with arguments. Maybe also `data_organizer.m`.

- What is the rationale for this computation https://github.com/VU-BEAM-Lab/GENRE/blob/master/GENRE_Code/GENRE.m#L111
