### 2020-09-12 VU-BEAM-Lab/GENRE#1 @marouenbg
- Is `GENRE_GPU_double_precision.cu` a standalone, i.e., can it be installed through a compiler without MATLAB? If so, I suggest adding the installation instructions and a minimal working example.

- I went through the cuda file and I think you did a flawless implementation with a very clear and easy-to-follow code, congratulations! I liked the part where you optimized the memory access through local and shared to further optimize the run time. It could be worthwhile to discuss this feature of the implementation (memory types and speed) in the manuscript.

- Memory is often the limiting resource on GPU, do you expect applications or model size that would require more than the capacity of current GPUs? Do you think that an implementation in "streams" like many machine learning packages have, would be address this type of issues?

- Final thought, what are the advantages of your implementation over coding it directly in MATALB using the parallel toolobx and the `gpuArray` routine?

These are my final thoughts. I think the implementation is very clean, with very clear code and docs. You went the extra mile to implement the software in cuda and used mex to make it callable from MATLAB. This is an excellent piece of software.
