### 2020-09-12 edbd8c3
Update GPU_kernels_double_precision.cu
