### 2020-09-12 4e2b344
Update GPU_kernels_single_precision.cu
