### 2020-09-12 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
Hi @marouenbg. Thank you for the additional comments about the code! It's great to have a fellow CUDA programmer approve of the CUDA implementation :). Also, I'm really glad to hear the code was easy to follow. I had spent a good amount of time trying to clean it up and comment it.

I've addressed some of your additional comments. I've added bullet points to the relevant sections, and I have added a Contributing section to the README.md file. My contribution guidelines are relatively simple, which is why I just added a section in the README.md instead of making a separate contribute.md file. I will work on addressing the rest of the comments in the next few days. I will add the benchmark in the manuscript. In terms of the test functionality, I actually haven't used TRAVIS-CI before, but I'll take a look at it. If I don't use that, I will just use your other suggestion and write a testGENRE() function that creates a small dataset and fits it like in my tutorial. I guess I can have the output just be a plot of the original y signal along with the y signal that the model fit produces. The user will be able to see if the model fit has actually been performed or not based off of this.

In terms of making GENRE a function instead of a script, I was actually hoping to get feedback on this from the review process, so thank you for bringing it up. I had just kept it as a script in case if MATLAB users wanted to access the workspace variables once the script finishes. So in your opinion, do you think a function format is more user-friendly and simpler? If so, do you think I should just include the model coefficients as the only function output? Also, I can make the data_organizer.m script be a function too, but it will just end up using arguments that you would pass into the GENRE function, which is why I had just kept it as a script.

In terms of the rationale for the computation that you had pointed out, in that section, I see if shared memory can be used. If so, I compute the maximum number of threads that can be used within one block. I then find the closest power of 2 that is below this number. Essentially, I'm just making sure the block size is a multiple of the warp size, which is 32 (some of the older GPUs have a warp size of 16). You're probably familiar with this already due to your experience using CUDA, but using a multiple of the warp size helps ensure that resources aren't being wasted. For example, suppose I have 32 model fits to perform, but I set 31 as the number of threads to use within the block. This will waste resources because two blocks will be needed. The first block will have a warp where 31 threads are doing work, and one thread is just masked out. The other block will just have 1 thread doing work with the others masked out. This is wasteful because you have to issue instructions to two warps in this case. Moreover, depending on the problem, you might be wasting memory bandwidth in this scenario if all 32 model fits need access to values that are stored in contiguous memory locations. This is because if you had used a block size of 32 instead of 31, then only one warp would be executed, and one coalesced memory transaction could be used to service all of the threads whenever a memory access occurs.

In terms of GENRE_GPU_double_precision.cu, it can't be used as a standalone as is, but it can be with modification. For example, if you just wanted to use the C/CUDA code without MATLAB, you would just need to get rid of the components in the file that are associated with the MEX-interface. These are mainly the parts where data is being passed back and forth with MATLAB (towards the beginning and end of the file). You would then need to determine how you want to access the data variables once the MEX-interface is removed. For example, you can just read from and write to binary files. The nice thing is that the lines of code that actually call the CUDA kernels and the CUDA kernels themselves would remain the same. In terms of compiling, you should just be able to compile it with nvcc. Now, if you want to use another high level language, then you would need to replace the MEX-interface with whatever interface the language uses. Languages such as R, Python, and Julia all have some way to call CUDA kernels. I didn't include any specific instructions because there are a variety of ways to potentially use it without MATLAB.

In terms of memory being a limit, I completely agree. Given the large spectrum of problem domains that this software can potentially be applied to, I think there will be situations where the GPU memory capacity will not be enough. I agree with your suggestion of stream processing as being a potential solution, especially if the size of each individual model fit is very large. With this approach though, you would also have to consider the performance penalty from having to perform multiple data transfers between host and device. In addition, in the case where you want to perform an extremely large number of model fits and can't fit all of the fits onto one GPU, then distributing the computations across multiple GPUs could also be another approach.

In terms of the advantages of a custom CUDA implementation vs MATLAB gpuArrays, my general opinion is that writing custom CUDA code has more potential to be tuned, especially if it is being tuned to a specific problem. For example, in this application, using shared memory is an example optimization that was used, and I'm not sure if the gpuArray approach would be able to perform the same optimization in this case. Therefore, due to this, I would expect the custom approach to have more performance benefits. I believe that the MATLAB gpuArrays will be most beneficial to people who do not necessarily want to dive deep into CUDA but would still like to achieve some additional performance benefits. The gpuArrays only support certain functions, so this may or may not be sufficient for a given application. Moreover, another benefit of using custom CUDA code for this application is that it makes it easier to use with languages other than MATLAB and as a standalone because you already have all of the source code. However, MATLAB has some tools to generate source code as well, but I haven't really used them. Adding on, I personally like custom CUDA code because it allows you to have more control with what's going on under the hood in terms of implementation and resource management.

This post was really long haha, but your comments and feedback have been very helpful in improving this repository!
