### 2020-09-12 VU-BEAM-Lab/GENRE#1 @marouenbg
Thank you for addressing all these questions in great detail.

> So in your opinion, do you think a function format is more user-friendly and simpler?

yes, I think it could help to have a `GENRE()` function.
