+++ README.md
@@ -262,7 +262,7 @@ Note that since we included an intercept term in every model, the first model co
    corresponding to the intercept term if it is included. This is due to the fact that standardization or normalization is applied to all of
    the other predictor columns. Therefore, a division by 0 will occur if all of the observations in a predictor column are the same.
    
-3. The ```GENRE.m``` script calls either the ```GENRE_GPU_single_precision``` MEX-file or the ```GENRE_GPU_double_precision``` MEX-file  
+3. The ```GENRE.m``` function calls either the ```GENRE_GPU_single_precision``` MEX-file or the ```GENRE_GPU_double_precision``` MEX-file  
    depending on what precision is specified. There is some additional overhead the first time either of these MEX-files is called. This is because on the
    first call, all of the input arrays generated by the ```data_organizer.m``` script are transferred to the GPU. For example the model
    matrices are transferred. However, upon subsequent calls to either MEX-file, these arrays do not need to be transferred again because 
