+++ README.md
@@ -266,7 +266,7 @@ Note that since we included an intercept term in every model, the first model co
    depending on what precision is specified. There is some additional overhead the first time either of these MEX-files is called. This is because on the
    first call, all of the input arrays generated by the ```data_organizer.m``` script are transferred to the GPU. For example the model
    matrices are transferred. However, upon subsequent calls to either MEX-file, these arrays do not need to be transferred again because 
-   they are already on the GPU. An example of when this might be relevant is if the ```GENRE.m ``` script is modified to call either of 
+   they are already on the GPU. An example of when this might be relevant is if the ```GENRE.m ``` function is modified to call either of 
    these MEX-files in a for loop. The only arrays that are transferred every time either MEX-file is called are ```alpha_values_h```, ```lambda_values_h```, ```tolerance_values_h```, ```max_iterations_values_h```, and ```y_h```. The first four arrays contain model fit
    parameters, and the last array is a 1-D array that contains the data contained in `y` for every model fit. This allows for functionality,
    such as calling one of the MEX-files in a for loop that produces the model coefficients for all of the model fits using
@@ -277,12 +277,12 @@ Note that since we included an intercept term in every model, the first model co
    called in order to save computations for subsequent calls. Note that the model matrices cannot be changed if the MEX-files are called
    within a for loop. If it is desired to change the model matrices, then ```clear mex``` must be called in order to unload any
    MEX-functions from memory and to free the memory allocated on the GPU. The ```GENRE.m``` function, as is, does not call the MEX-files in a 
-   for loop, so ```clear mex``` is included at the top of the script. Therefore, each time you run the script, it will be called before 
+   for loop, so ```clear mex``` is included at the top of the function. Therefore, each time you run the function, it will be called before 
    either of the MEX-files is called. 
    
 4. As previously stated, each model fit is supplied a lambda value in ```GENRE```, so the coefficients are not calculated for a path of 
    lambda values. One way of doing this in ```GENRE``` is to make multiple model files containing the same model. Then, a different lambda 
-   value can be supplied to each model fit via the ```lambda_values_h``` vector in the ```GENRE.m``` script. Therefore, the model 
+   value can be supplied to each model fit via the ```lambda_values_h``` vector that is an input to the ```GENRE.m``` function. Therefore, the model 
    coefficients for the same model using different lambda values will be computed in parallel on the GPU. In addition, another way of 
    calculating the coefficients for multiple lambda values is to follow the method described in 3 above. For example, the ```GENRE.m``` 
    function can be modified to call either the ```GENRE_GPU_single_precision``` MEX-file or the ```GENRE_GPU_double_precision``` MEX-file in a 
