+++ README.md
@@ -266,7 +266,7 @@ Note that since we included an intercept term in every model, the first model co
    depending on what precision is specified. There is some additional overhead the first time either of these MEX-files is called. This is because on the
    first call, all of the input arrays generated by the ```data_organizer.m``` script are transferred to the GPU. For example the model
    matrices are transferred. However, upon subsequent calls to either MEX-file, these arrays do not need to be transferred again because 
-   they are already on the GPU. An example of when this might be relevant is if the ```GENRE.m ``` function is modified to call either of 
+   they are already on the GPU. An example of when this might be relevant is if the ```GENRE.m``` function is modified to call either of 
    these MEX-files in a for loop. The only arrays that are transferred every time either MEX-file is called are ```alpha_values_h```, ```lambda_values_h```, ```tolerance_values_h```, ```max_iterations_values_h```, and ```y_h```. The first four arrays contain model fit
    parameters, and the last array is a 1-D array that contains the data contained in `y` for every model fit. This allows for functionality,
    such as calling one of the MEX-files in a for loop that produces the model coefficients for all of the model fits using
