### 2020-09-13 3d02d9b
Changed GENRE.m script to a function
