### 2020-09-13 c1d236f
Added a test function
