+++ README.md
@@ -67,11 +67,20 @@ mexcuda GENRE_GPU_double_precision.cu -L/usr/local/cuda-10.0/lib64
 ```
 
 ### Testing GENRE
-* To ensure that the setup process is successful, the ```test_GENRE.m``` function is included in the ```GENRE_Code``` folder. Once the previous steps have been completed, type the command shown below into the MATLAB command prompt. This function generates a test dataset consisting of model matrices and data to fit the models to. This dataset is stored into a folder called ```Test_Models```, which is created when the function is executed. The function then performs the model fits using ```GENRE```. Both double precision and single precision are tested for the model fits. If the tests are successful, then a folder called ```Test_Model_Coefficients``` will be created, and it will contain two ```.mat``` files called ```model_coefficients_single_precision.mat``` and ```model_coefficients_double_precision.mat```.
+* To ensure that the setup process is successful, the ```test_GENRE.m``` function is included in the ```GENRE_Code``` folder. Once the previous steps have been completed, type the command shown below into the MATLAB command prompt. 
 
 ```Matlab
 test_GENRE
 ```
+
+* This function generates a test dataset consisting of model matrices and data to fit the models to. This dataset is stored into a folder called ```Test_Models```, which is created when the function is executed. The function then performs the model fits using ```GENRE```. Both double precision and single precision are tested for the model fits. If the tests are successful, then a folder called ```Test_Model_Coefficients``` will be created, and it will contain two ```.mat``` files called ```model_coefficients_single_precision.mat``` and ```model_coefficients_double_precision.mat```. These files contain the computed model coefficients for the model fits and the parameters that were used for the model fits. In particular, the computed model coefficients are contained in the variable ```B_cell```, where each entry in this cell corresponds to a different model fit. For example, to see the result of the first model fit, the user can load in the ```model_data_1.mat``` file located in the ```Test_Models``` folder. The ```X``` variable that is contained in this file corresponds to the model matrix for the first model fit, and the ```y``` variable corresponds to the data that the model matrix was fit to. The data that is reconstructed by the first model fit can be obtained and compared to ```y``` by typing the command shown below in the MATLAB command prompt.
+
+```Matlab
+y_first_model_fit = X * B_cell{1}
+figure;
+plot(y, 'b');
+hold on;
+plot(y_first_model_fit, 'r');
 
 ## Model Data Format
 As previously stated, ```GENRE``` allows for many models to run in parallel on the GPU. The data for each model fit needs to be saved as a ```.mat``` file. For example, if there are 100 model fits that need to be performed, then there should be 100 ```.mat``` files. Each file should contain the following 3 variables. 
