+++ README.md
@@ -73,7 +73,7 @@ mexcuda GENRE_GPU_double_precision.cu -L/usr/local/cuda-10.0/lib64
 test_GENRE
 ```
 
-* This function generates a test dataset consisting of model matrices and data to fit the models to. This dataset is stored into a folder called ```Test_Models```, which is created when the function is executed. The function then performs the model fits using ```GENRE```. Both double precision and single precision are tested for the model fits. If the tests are successful, then a folder called ```Test_Model_Coefficients``` will be created, and it will contain two ```.mat``` files called ```model_coefficients_single_precision.mat``` and ```model_coefficients_double_precision.mat```. These files contain the computed model coefficients for the model fits and the parameters that were used for the model fits. In particular, the computed model coefficients are contained in the variable ```B_cell```, where each entry in this cell corresponds to a different model fit. For example, to see the result of the first model fit, the user can load in the ```model_data_1.mat``` file located in the ```Test_Models``` folder. The ```X``` variable that is contained in this file corresponds to the model matrix for the first model fit, and the ```y``` variable corresponds to the data that the model matrix was fit to. The data that is reconstructed by the first model fit can be obtained and compared to ```y``` by typing the command shown below in the MATLAB command prompt.
+* This function generates a test dataset consisting of model matrices and data to fit the models to. This dataset is stored into a folder called ```Test_Models```, which is created when the function is executed. The function then performs the model fits using ```GENRE```. Both double precision and single precision are tested for the model fits. If the tests are successful, then a folder called ```Test_Model_Coefficients``` will be created, and it will contain two ```.mat``` files that are called ```model_coefficients_single_precision.mat``` and ```model_coefficients_double_precision.mat```. These files contain the computed model coefficients for the model fits and the parameters that were used for the model fits. In particular, the computed model coefficients are contained in the variable ```B_cell```, where each entry in this cell corresponds to a different model fit. For example, to see the result of the first model fit, the user can load in the ```model_data_1.mat``` file located in the ```Test_Models``` folder. The ```X``` variable that is contained in this file corresponds to the model matrix for the first model fit, and the ```y``` variable corresponds to the data that the model matrix was fit to. The data that is reconstructed by the first model fit can be obtained and compared to ```y``` by typing the command shown below in the MATLAB command prompt.
 
 ```Matlab
 y_first_model_fit = X * B_cell{1}
