### 2020-09-13 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
Hi @marouenbg. I have converted GENRE.m from a script to a function. In addition, I have added a test_GENRE.m function, and I have updated the Setup section in the README.md file to explain it. Is this sufficient for the testing component from your perspective?

Thank you for your time.
