### 2020-09-13 a84a8b1
Added test_GENRE_shared_memory.m function
