### 2020-09-13 e7e179b
Edited test_GENRE_shared_memory.m
