### 2020-09-13 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
In addition to my above comment, I have also added a test_GENRE_shared_memory.m function. It does the same thing as the test_GENRE.m function, but it uses smaller model matrix sizes to test for the case where shared memory is used in addition to global memory for the model fits. I have described it in the Setup section as well.
