### 2020-09-13 VU-BEAM-Lab/GENRE#1 @marouenbg
One last thing, if you're going to keep the speed up claims in the paper, they have to be backed up by benchmark data of GPU vs CPU. Also add the CPU and GPU hardware specifications that you're testing on.
