+++ Paper/paper.md
@@ -53,7 +53,7 @@ another interface if it is desired to call the C/CUDA code in another language, 
 without an interface.
 
 # Statement of Need
-The core motivation for developing GENRE was that many of the available packages for performing linear regression with elastic-net regularization focus on achieving high performance in terms of computational time or resource consumption for single model fits. However, they often do not address the case in which there is a need to perform many model fits in parallel. For example, the research project that laid the foundation for GENRE involved performing ultrasound image reconstruction using an algorithm called Aperture Domain Model Image REconstruction (ADMIRE) [@byram_jakovljevic_2014; @byram_dei_tierney_dumont_2015; @dei_byram_2017]. This algorithm is computationally expensive due to the fact that in one stage, it requires thousands of instances of linear regression with elastic-net regularization to be performed in order to fit models of ultrasound data. When this algorithm was implemented on a CPU, it typically required an amount of time that was on the scale of minutes to reconstruct one ultrasound image. The primary bottleneck was performing all of the required model fits due to the fact that a custom C implementation of cyclic coordinate descent was used to compute each fit serially. However, a GPU implementation of the algorithm was developed, and this implementation provided a speedup of over two orders of magnitude, which allowed for multiple ultrasound images to be reconstructed per second. The main contributor to this speedup was the fact that the model fits were performed in parallel on the GPU. For example, on a computer containing dual Intel Xeon E5-2650 v2 CPUs @ 2.60 GHz with 8 cores each along with an NVIDIA GeForce GTX 1080 Ti GPU and an NVIDIA GeForce RTX 2080 Ti GPU, the CPU implementation of ADMIRE had an average processing time of 90.033 $\pm$ 0.494 seconds to process one frame of ultrasound channel data while the GPU implementation had an average processing time of 0.433 $\pm$ 0.001 seconds. The average processing time was obtained by taking the average of 10 runs for the same dataset. Note that the 2080 Ti GPU was used to perform GPU processing.
[…]
