### 2020-09-14 bef3c72
Update paper.md
