### 2020-09-14 cb9c19d
Update paper.md
