### 2020-09-14 defd3f4
Update paper.md
