### 2020-09-15 6d3d6ce
Update paper.md
