### 2020-09-15 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
Hi @marouenbg. Sorry for the late response. I have added the ADMIRE benchmark to the paper. One thing that I updated is that we had originally used glmnet for the CPU implementation, but I recently wrote a custom C implementation of cyclic coordinate descent that is faster. Therefore, we use this for the CPU implementation of ADMIRE now, and I have updated the paper to reflect this. In addition, I liked your idea of adding an extra section to the paper demonstrating the performance comparison when using shared memory and when not using shared memory. I can also add a benchmark comparing GENRE to glmnet in this section. I will try and do this by the end of this week.

Thank you for your time.
