### 2020-09-15 0d270a6
Update paper.md
