### 2020-09-16 96c4091
Update paper.md
