+++ Paper/paper.md
@@ -61,12 +61,12 @@ Aside from this application, there are a number of other applications that can p
 GENRE has the potential to provide significant speedup due to the fact that many model fits can be performed in parallel on a GPU. Therefore, an example benchmark was performed where we compared GENRE to glmnet, which is written in Fortran and performs the model fits in a serial fashion on a CPU. In this benchmark, 20,000 model matrices were randomly generated within MATLAB. Each model matrix consisted of 50 observations and 200 predictors (50x200). For each model matrix, the model coefficients were randomly generated, and the matrix multiplication of the model matrix and the coefficients was performed to obtain the observation vector. Therefore, this provided 20,000 observation vectors with each containing 50 observations. Once the data was generated, both GENRE and glmnet were used to perform the model fits and return the computed model coefficients. GENRE allows for the user to select either single precision or double precision for performing the model fits on the GPU, so processing was done for both cases. The MATLAB version of the glmnet software package includes a compiled executable MEX-file that allows for Fortran code to be called, and it uses double precision for the calculations. In addition, due to the fact that all of the model matrices have a small number of observations (50) in this case, GENRE is also able to use shared memory in addition to global memory when performing the model fits. Shared memory has lower latency than global memory, so utilizing it can provide performance benefits. Therefore, processing was performed both with and without using shared memory. The computer that was used for the benchmarks contained dual Intel Xeon Silver 4114 CPUs @ 2.20 GHz with 10 cores each along with an NVIDIA GeForce GTX 1080 Ti GPU and an NVIDIA GeForce RTX 2080 Ti GPU. The 2080 Ti GPU was used to perform GPU processing. For each case, the average of 10 runs was taken, and timing was performed using MATLAB's built-in timing capabilities. Note that GENRE has a data organization step that loads data from files and organizes it in the format that is used by the GPU. For this benchmark, this step was not counted in the timing due to the fact that it was assumed that all of the data was already loaded into MATLAB for both GENRE and glmnet. In addition, the GPU times also include the time it takes to transfer data for the model fits from the host system to the GPU and the time it takes to transfer the computed model coefficients back from the GPU to the host system. The benchmark results are shown in Table I below.
 
  \begin{centering}
- \scriptsize
+ \footnotesize
 \begin{tabular}[h]{| c | c | c | c | c |} 
 \hline
 \multicolumn{5}{|c|}{\textbf{TABLE I: Benchmark Times (s)}} \\
 \hline
-\textbf{GENRE Double Precision} & \textbf{GENRE Double Precision SMEM} & \textbf{GENRE Single Precision} & \textbf{GENRE Single Precision SMEM} & \textbf{glmnet} \\
+\textbf{GENRE DP} & \textbf{GENRE DP SMEM} & \textbf{GENRE SP} & \textbf{GENRE SP SMEM} & \textbf{glmnet} \\
 \hline
  1.378 $\pm$ 0.016 & 1.152 $\pm$ 0.020 & 1.083 $\pm$ 0.021 & 0.943 $\pm$ 0.027 & 10.542 $\pm$ 0.136 \\
 \hline
