### 2020-09-16 cbf6648
Update paper.md
