+++ Paper/paper.md
@@ -96,7 +96,7 @@ As shown in Table I, GENRE provides an order of magnitude speedup when compared 
 
 \vspace{0.3 cm}
 
-As shown in Table II, when the model matrices can be reused and do not have to be reloaded, GENRE provides a speedup of over two orders of magnitude when compared with glmnet, and using single precision with shared memory provides the best performance. This type of performance gain would most likely be difficult to achieve even when using a multi-CPU implementation of cyclic coordinate descent on a single host system.
+As shown in Table II, when the model matrices can be reused and do not have to be reloaded, GENRE provides a speedup of over two orders of magnitude when compared with glmnet, and using single precision with shared memory provides the best performance. This type of performance gain would most likely be difficult to achieve even when using a multi-CPU implementation of cyclic coordinate descent on a single host system. In addition, it is important to note that this benchmark was just to illustrate an example of when using GENRE provides performance benefits, but it does not mean that it will always outperform a CPU implementation. For example, in GENRE, one computational thread on the GPU is used to perform each model fit. Therefore, when many model fits need to be performed, the parallelism of the GPU can be utilized. However, if only one fit needs to be performed, for example, then using a serial CPU implementation such as glmnet will most likely provide better performance due to factors such as CPU cores having higher clock rates and more resources per core than GPU cores.
 
 # Acknowledgements
 This work was supported by NIH grants R01EB020040 and S10OD016216-01 and NAVSEA grant N0002419C4302.
