### 2020-09-16 8ca6f46
Update paper.md
