### 2020-09-16 4ade8a9
Update paper.md
