### 2020-09-16 c297fbe
Update paper.md
