### 2020-09-16 1c396ea
Update paper.md
