### 2020-09-16 d17e42e
Update paper.md
