+++ Paper/paper.md
@@ -79,7 +79,7 @@ The computer that was used for the benchmarks contained dual Intel Xeon Silver 4
 
 \vspace{0.3 cm}
 
-As shown in Table I, GENRE provides an order of magnitude speedup when compared to glmnet, and the best performance was achieved by using single precision with shared memory. In addition, it is important to note that in these benchmarks, most of the time for GENRE was spent transferring the model matrices from the host system to the GPU. However, there are cases when once the model matrices have been used in one call, they can be reused in subsequent calls. For example, a user might want to reuse the same model matrices except just change the $\alpha$ value or the $\lambda$ value that is used in elastic-net regularization, or they might want to just change the observation vectors that the model matrices are fit to. By default, each time GENRE is called, the \textit{clear mex} command is executed, and the GENRE MEX-files are setup so that all allocated memory on the GPU is freed when this command is called. However, in a case where the model matrices can be reused after they are loaded once, the \textit{clear mex} command can be removed. Essentially, every time one of the MEX-files for GENRE is called for the first time, all of the data for the model fits will be loaded onto the GPU. However, for subsequent calls without the \textit{clear mex} command, all of the data for the model fits will still be reloaded except for the model matrices, which will be kept on the GPU from the first call. By not having to reload the model matrices, performance can be significantly increased. To demonstrate this, the same benchmark from above was repeated, but for each case this time, GENRE was called once before performing the 10 runs. This is to replicate the case where the model matrices are reused in subsequent calls. The benchmark results are shown in Table II below. Note that the benchmark for glmnet was not repeated.
+As shown in Table I, GENRE provides an order of magnitude speedup when compared to glmnet, and the best performance was achieved by using single precision with shared memory. In addition, it is important to note that in these benchmarks, most of the time for GENRE was spent transferring the model matrices from the host system to the GPU. However, there are cases when once the model matrices have been used in one call, they can be reused in subsequent calls. For example, a user might want to reuse the same model matrices except just change the $\alpha$ value or the $\lambda$ value that is used in elastic-net regularization, or they might want to just change the observation vectors that the model matrices are fit to. By default, each time GENRE is called, the \textit{clear mex} command is executed, and the GENRE MEX-files are setup so that all allocated memory on the GPU is freed when this command is called. However, in a case where the model matrices can be reused after they are loaded once, the \textit{clear mex} command can be removed. Essentially, every time one of the MEX-files for GENRE is called for the first time, all of the data for the model fits will be loaded onto the GPU. However, if the \textit{clear mex} command is removed, then for subsequent calls, all of the data for the model fits will still be reloaded except for the model matrices, which will be kept on the GPU from the first call. By not having to reload the model matrices, performance can be significantly increased. To demonstrate this, the same benchmark from above was repeated, but for each case this time, GENRE was called once before performing the 10 runs. This is to replicate the case where the model matrices are reused in subsequent calls. The benchmark results are shown in Table II below. Note that the benchmark for glmnet was not repeated.
 
 \vspace{0.3 cm}
 
