### 2020-09-16 fa7311f
Update paper.md
