### 2020-09-16 603e6ac
Update paper.md
