### 2020-09-16 VU-BEAM-Lab/GENRE#1 @Christopher-Khan
Hi @marouenbg. I have updated the manuscript with the section containing the benchmark results, so it should be finalized now.

Thank you for your time.
Christopher Khan
