### 2020-09-17 #2644 @Christopher-Khan
Hi @marouenbg. I just added one more result to the manuscript. The glmnet software package has two algorithm options to choose from, so I also included the benchmark result for the other option as well. It was slower than the option whose benchmark result I had already reported though. In addition, thank you again for doing this review! Your comments were very helpful in improving this submission!
