+++ Paper/paper.bib
@@ -33,6 +33,12 @@
         url={http://www.web.stanford.edu/~hastie/Papers/Glmnet_Vignette.pdf}
 }
 
+@article{H2O4GPU_2020,
+	title={H2O4GPU}, 
+	year={2020},
+	url={https://github.com/h2oai/h2o4gpu}
+}	
+
 @article{byram_jakovljevic_2014, 
 	title={Ultrasonic multipath and beamforming clutter reduction: a chirp model approach}, 
 	volume={61},
