### 2020-09-22 VU-BEAM-Lab/GENRE#3 @krystophny
Testing on Linux
Two minor things happened to me

1) One has to add the GENRE_Code to the path together with subfolders for tests to work, please update this in the docs
2) The extra '-L/usr/local/cuda-10.0/lib64' for Linux compilation was not required in my case, anyway one should mention to replace this by the actual CUDA path in case (mine is 10.2)
