### 2020-09-22 #2644 @krystophny
@sjpfenninger I have finished reviewing the paper and testing the software and support the publication in JOSS after two minor changes in the README https://github.com/VU-BEAM-Lab/GENRE/issues/3 . As mentioned the package is relatively small (1000-2000 lines of code) and not very well known yet, but it looks very useful with its unique capability to solve multiple GLMs with elastic net regularization in parallel on GPUs and has the potential as a standard code for that task. The description is done well, and comparable software (without the special parallel feature) is referenced.
