### 2020-09-22 #2644 @Christopher-Khan
Hi @krystophny. Thank you for your review! In regards to the minor points, my responses are below.

For adding the folders to the MATLAB path, in the bullet point under Windows and Linux installation in the README, I have the following statement, "Assuming the repository is on your system, go to the MATLAB directory that contains the repository folders and add them to your MATLAB path."

In regards to not needing to specify the cuda path for Linux, I have the following statement in the second bullet point under Linux compilation in the README, "Note that mexcuda might find the CUDA toolkit library even if you do not explicitly type out its path. In addition, note that there might be differences in your path compared to the one shown above, such as in regards to the version of the CUDA toolkit that is being used."

Did you want me to still modify those statements?
