### 2020-09-23 #2644 @sjpfenninger
@marouenbg @krystophny Thanks both for your reviews. Can you both confirm that you recommend to accept the submission? @krystophny can you close VU-BEAM-Lab/GENRE#3 if it is resolved?
