### 2020-09-24 bac0710
Update paper.md
