+++ Paper/paper.md
@@ -69,16 +69,17 @@ The computer that was used for the benchmarks contained dual Intel Xeon Silver 4
 \vspace{0.3 cm}
 
  \begin{centering}
+ \begin{table}
+ \caption{\label{tab:table-1}Benchmark times (seconds).}
  \footnotesize
 \begin{tabular}[h]{| c | c | c | c | c |} 
-\hline
-\multicolumn{5}{|c|}{\textbf{TABLE I: Benchmark Times (s)}} \\
 \hline
 \textbf{GENRE DP} & \textbf{GENRE DP SMEM} & \textbf{GENRE SP} & \textbf{GENRE SP SMEM} & \textbf{glmnet} \\
 \hline
  1.367 $\pm$ 0.024 & 1.137 $\pm$ 0.015 & 1.048 $\pm$ 0.008 & 0.909 $\pm$ 0.006 & 10.319 $\pm$ 0.039 \\
 \hline
 \end{tabular}
+\end{table}
 \end{centering}
 
 \vspace{0.3 cm}
