### 2020-09-24 a2f6d4f
Update paper.md
