+++ Paper/paper.md
@@ -45,7 +45,7 @@ Cycling through all of the coefficients results in one iteration, and this proce
 convergence criteria are satisfied. As previously stated, glmnet is highly efficient for single model fits, but
 performing thousands of these fits will still require significant computational time due to each one being executed 
 in a serial fashion on a CPU. However, by using GENRE, massively parallel processing can be performed in order to 
-achieve significant speedup. This is due to the fact that modern GPUs consist of thousands of computational cores
+achieve a significant speedup. This is due to the fact that modern GPUs consist of thousands of computational cores
 that can be utilized. Moreover, although the processing in GENRE is performed using the C programming language and 
 CUDA, a MEX-interface is included to allow for this code to be called within the MATLAB (The MathWorks, Inc., Natick, MA, USA) 
 programming language for convenience. This also means that with modification, the MEX-interface can be replaced with 
@@ -64,7 +64,7 @@ Aside from this application, there are a number of other applications that can p
 # Example Benchmark Comparing GENRE with glmnet
 GENRE has the potential to provide significant speedup due to the fact that many model fits can be performed in parallel on a GPU. Therefore, an example benchmark was performed where we compared GENRE with glmnet, which is written in Fortran and performs the model fits in a serial fashion on a CPU. In this benchmark, 20,000 model matrices were randomly generated within MATLAB. Each model matrix consisted of 50 observations and 200 predictors (50x200), and an intercept term was included for all of the models. Note that to add an intercept term in GENRE, a column of ones was appended at the beginning of each model matrix to make the predictor dimension 201 (adding a column of ones is not required for glmnet). For each model matrix, the model coefficients were randomly generated, and the matrix multiplication of the model matrix and the coefficients was performed to obtain the observation vector. Therefore, this provided 20,000 observation vectors with each containing 50 observations. Once the data was generated, both GENRE and glmnet were used to perform the model fits and return the computed model coefficients. An $\alpha$ value of 0.5 and a $\lambda$ value of 0.001 were used for all of the model fits. The tolerance convergence criterion for both packages was set to 1E-4. It was also specified for each package to standardize the model matrices, which means that the unstandardized model coefficients were returned. Note that the column of ones for each model matrix corresponding to the intercept term is not standardized in the case of GENRE. GENRE allows for the user to select either single precision or double precision for performing the model fits on the GPU, so processing was done for both cases. The MATLAB version of the glmnet software package includes a compiled executable MEX-file that allows for Fortran code to be called, and it uses double precision for the calculations. In addition, due to the fact that all of the model matrices have a small number of observations (50) in this case, GENRE is also able to use shared memory in addition to global memory when performing the model fits. Shared memory has lower latency than global memory, so utilizing it can provide performance benefits. Therefore, processing was performed both with and without using shared memory. 
 
[…]
