### 2020-09-24 #2644 @sjpfenninger
@Christopher-Khan I have made some minor edits to the paper in VU-BEAM-Lab/GENRE#4 -- can you check and merge (ideally squash into single commit)? Also - could you adjust the table captions such that they differentiate what we see in the two tables rather than just saying "Benchmark times"?
