### 2020-09-24 6ae9f8c
Minor edits to JOSS paper (#4)
* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md

* Update paper.md
