### 2020-09-24 c744360
Update paper.md
