### 2020-09-28 #2644 @sjpfenninger
@Christopher-Khan great thanks. Can you edit the Zenodo item such that title and author list are the same as for the JOSS paper?
