### 2020-09-29 #2644 @Christopher-Khan
Hi @sjpfenninger. Sorry for the late response. I made one more edit to the paper. I forgot for the benchmarks to account for the fact that MATLAB code can sometimes run slower the first time it is called than subsequent calls due to MATLAB's execution engine. Therefore, I reran and updated the benchmarks in order to account for this. The times didn't change much, but I still wanted to update them in the paper. I made a new release on Zenodo and detailed the change on there. I have also updated the metadata. Here is the new Zenodo DOI.  DOI: 10.5281/zenodo.4057940

Thank you!
