### 2020-10-09 #2644 @kthyng
Hi @Christopher-Khan! I see that you archive and version are set and up to date. I just looked at your paper and unfortunately it is far too long. A [JOSS paper](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) is to be between 250 and 1000 words. Can you move some of the content from the paper to your documentation?
