### 2020-10-09 #2644 @Christopher-Khan
Hi @kthyng! Sure. I can move the section titled "Example Benchmark Comparing GENRE with glmnet" to my documentation on GitHub. That will make the paper two pages shorter. Do you want me to make a new Zenodo archive then and give the DOI for that one? I ask because I don't think Zenodo allows for files in archives to be updated if the archive is more than a week old. Thank you!
