### 2020-10-10 #2644 @krystophny
@Christopher-Khan if you make a new release on github, Zenodo will create a new version of the same entry with a new doi. Probably that’s better than creating a new archive
