+++ README.md
@@ -332,23 +332,7 @@ If you want to compare ```GENRE``` to other packages that perform linear regress
 
 The computer that was used for the benchmarks contained dual Intel Xeon Silver 4114 CPUs @ 2.20 GHz with 10 cores each along with an NVIDIA GeForce GTX 1080 Ti GPU and an NVIDIA GeForce RTX 2080 Ti GPU. The 2080 Ti GPU was used to perform GPU processing. For each case, the average of 10 runs was taken, and timing was performed using MATLAB's built-in timing capabilities. Note that GENRE has a data organization step that loads the data for the model fits from files and organizes it into the format that is used by the GPU. For this benchmark, this step was not counted in the timing due to the fact that it was assumed that all of the data was already loaded into MATLAB on the host system for both ```GENRE``` and ```glmnet```. The GPU times include the time it takes to transfer data for the model fits from the host system to the GPU, standardize the model matrices, perform the model fits, unstandardize the model coefficients, transfer the computed model coefficients back from the GPU to the host system, and store the coefficients into a MATLAB cell structure. The CPU time includes the time it takes to standardize the model matrices, perform the model fits, unstandardize the model coefficients, and store the coefficients into a MATLAB cell structure. The benchmark results are shown in Table 1 below. Note that DP, SP, and SMEM correspond to double precision, single precision, and shared memory respectively. In addition, note that the input data for the model fits was of type `double` for this benchmark. Therefore, in the case of ```GENRE```, some of the inputs would need to be converted to type `single` before they are passed to the GPU when using single precision for the computations. Moreover, ```GENRE``` also converts the datatype of the computed model coefficients to the datatype of the original input data. This means that for the single precision cases, the computed model coefficients would need to be converted to be type `double` after they are passed back to the host system from the GPU. For purposes of benchmarking the single precision cases, the time to perform the type conversions of the inputs to type `single` was not included, and the returned model coefficients were just kept as type `single`. This is due to the fact that including these times would increase the benchmark times for the single precision cases in this scenario, and if it were a different scenario, the double precision cases could be impacted instead of the single precision cases. For example, if the type of the original input data was `single` and double precision was used for the calculations, then these datatype conversions would have to be made for the double precision cases, but they would not have to be made for the single precision cases.
 
-\vspace{0.3 cm}
 
- \begin{centering}
- \begin{table}[h!]
- \caption{\label{tab:table1}Benchmark times (seconds).}
- \footnotesize
-\begin{tabular}[h]{| c | c | c | c | c |} 
-\hline
-\textbf{GENRE DP} & \textbf{GENRE DP SMEM} & \textbf{GENRE SP} & \textbf{GENRE SP SMEM} & \textbf{glmnet} \\
-\hline
- 1.339 $\pm$ 0.036 & 1.090 $\pm$ 0.028 & 1.035 $\pm$ 0.046 & 0.885 $\pm$ 0.029 & 9.809 $\pm$ 0.196 \\
-\hline
-\end{tabular}
-\end{table}
-\end{centering}
-
-\vspace{0.3 cm}
 
[…]
