### 2020-10-10 #2644 @Christopher-Khan
@krystophny thank you. That's what I meant actually instead of creating a new archive. @kthyng I have moved the section in the JOSS paper titled "Example Benchmark Comparing GENRE with glmnet" from the paper to the section in the README.md file titled "Comparing with Other Packages". The new DOI is below. I hope the paper is better now in terms of its length.

DOI: 10.5281/zenodo.4076520

Thank you for your time.
