### 2020-10-10 #2644 @Christopher-Khan
@sjpfenninger. Last thing. Do I need to set the version at the top of this review from v1.0.1 to v1.0.2 because that's the newest version that actually got accepted? If so, I believe only editors can do that. Thank you!
